Renamed models folders
