+++ Backend/CarRentalApp/CarRentalApp/Controllers/UsersController.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Immutable;
-using CarRentalApp.Models.BLL;
-using CarRentalApp.Models.DAL;
-using CarRentalApp.Models.WEB.Requests;
+using CarRentalApp.Models.Bll;
+using CarRentalApp.Models.Dal;
+using CarRentalApp.Models.Web.Requests;
 using CarRentalApp.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
