+++ Backend/CarRentalApp/CarRentalApp/Models/Bll/LoginModel.cs
@@ -1,4 +1,4 @@
-﻿namespace CarRentalApp.Models.BLL
+﻿namespace CarRentalApp.Models.Bll
 {
     public class LoginModel
     {
