+++ Backend/CarRentalApp/CarRentalApp/Models/Dal/User.cs
@@ -1,4 +1,4 @@
-﻿namespace CarRentalApp.Models.DAL
+﻿namespace CarRentalApp.Models.Dal
 {
     public class User
     {
