+++ Backend/CarRentalApp/CarRentalApp/Models/Web/Requests/RegistrationRequestBase.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using CarRentalApp.Validation;
 
-namespace CarRentalApp.Models.WEB.Requests
+namespace CarRentalApp.Models.Web.Requests
 {
     public abstract class RegistrationRequestBase
     {
