+++ Backend/CarRentalApp/CarRentalApp/Models/Web/Responses/AuthenticationResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace CarRentalApp.Models.WEB.Responses
+﻿namespace CarRentalApp.Models.Web.Responses
 {
     public class AuthenticationResponse
     {
