+++ Backend/CarRentalApp/CarRentalApp/Services/TokenService.cs
@@ -6,9 +6,8 @@ using CarRentalApp.Configuration.JWT.Access;
 using CarRentalApp.Configuration.JWT.Refresh;
 using CarRentalApp.Contexts;
 using CarRentalApp.Exceptions;
-using CarRentalApp.Models.BLL;
-using CarRentalApp.Models.DAL;
-using CarRentalApp.Models.WEB.Responses;
+using CarRentalApp.Models.Bll;
+using CarRentalApp.Models.Dal;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
